change ID columns to bigint

+++ internal/db/functions.go
@@ -19,7 +19,7 @@ func CreateOrderBook (db *gorm.DB, market BinanceMarket) BinanceOrderBook{
 
 
 func AddOrder(db *gorm.DB, rate float64, quantity float64, time time.Time, orderbook BinanceOrderBook) BinanceOrder{
-	order := BinanceOrder{0,  time, orderbook.ID, rate, quantity,}
+	order := BinanceOrder{0,  time, uint64(orderbook.ID), rate, quantity,}
 	err := db.Create(&order).Error
 	if err != nil{
 		log.Panic(err)
